test: move test helpers out of test folder
Otherwise jest will incorrectly recognise it as a test and fail
