+++ packages/forms/src/validators/__tests__/createCharacterLengthRange.test.js
@@ -1,5 +1,5 @@
 import { createCharacterLengthRange } from '../createCharacterLengthRange.js'
-import { testValidatorValues, allowsEmptyValues } from './helpers/index.js'
+import { testValidatorValues, allowsEmptyValues } from '../test-helpers'
 import { requiredArgumentErrorMessage } from '../helpers/index.js'
 
 describe('validator: createCharacterLengthRange', () => {
