+++ packages/forms/src/validators/__tests__/createMaxNumber.test.js
@@ -1,5 +1,5 @@
 import { createMaxNumber } from '../createMaxNumber.js'
-import { testValidatorValues } from './helpers/index.js'
+import { testValidatorValues } from '../test-helpers'
 
 describe('validator: createMaxNumber', () => {
     const maxSix = createMaxNumber(6)
