+++ packages/forms/src/validators/__tests__/email.test.js
@@ -1,5 +1,5 @@
 import { email, invalidEmailMessage } from '../email.js'
-import { testValidatorValues, allowsEmptyValues } from './helpers/index.js'
+import { testValidatorValues, allowsEmptyValues } from '../test-helpers'
 
 /*
  * A comprehensive list technically valid and invalid email addresses was
