+++ packages/forms/src/validators/__tests__/hasValue.test.js
@@ -1,5 +1,5 @@
 import { hasValue, hasValueMessage } from '../hasValue.js'
-import { testValidatorValues } from './helpers/index.js'
+import { testValidatorValues } from '../test-helpers'
 
 describe('validator: hasValue', () => {
     describe('should return undefined for allowed values', () => {
